final_project/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
colorama==0.4.6
contourpy==1.1.0
coverage==7.2.3
cycler==0.11.0
dj-database-url==2.0.0
Django==4.2
django-bootstrap-v5==1.0.11
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fonttools==4.40.0
importlib-resources==5.12.0
iniconfig==2.0.0
kiwisolver==1.4.4
lxml==4.9.2
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
Pillow==10.0.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pyparsing==3.1.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.6.3
tzdata==2023.3
zipp==3.15.0
